moves players into separate folder
