+++ lib/players/computer_player.rb
