+++ lib/players/human_console_player.rb
