+++ lib/players/human_web_player.rb
