+++ lib/players/random_player.rb
