+++ lib/web_controller.rb
@@ -1,6 +1,6 @@
 require 'sinatra'
 require 'game'
-require 'player_factory'
+require 'players/player_factory'
 require 'board_factory'
 require 'marks'
 require 'game_options'
